Remove python-docker
Installation was failing with pip
× This environment is externally managed

And with pipx
No apps associated with package docker. Try again with '--include-deps' to include apps of dependent packages, which are listed above. If you are attempting to install a library, pipx
should not be used. Consider using pip or a similar tool instead."

I don't need python docker now so removing it instead of making effort to fix